+++ tests/codes/eurocode/en_1993_1_1_2022/chapter_8_ultimate_limit_state/test_formula_8_51aw.py
@@ -1,8 +1,8 @@
-"""Testing formula 8.51aw from EN 1993-1-1:2025, chapter 8, ultimate limit state."""
+"""Testing formula 8.51aw from EN 1993-1-1:2022, chapter 8, ultimate limit state."""
 
 import pytest
 
-from blueprints.codes.eurocode.en_1993_1_1_2025.chapter_8_ultimate_limit_state.formula_8_51aw import (
+from blueprints.codes.eurocode.en_1993_1_1_2022.chapter_8_ultimate_limit_state.formula_8_51aw import (
     Form8Dot51awHollowSections,
     Form8Dot51awWeldedBoxSections,
 )
@@ -10,7 +10,7 @@ from blueprints.validations import LessOrEqualToZeroError, NegativeValueError
 
 
 class TestForm8Dot51awHollowSections:
-    """Validation for formula 8.51aw (hollow sections) from EN 1993-1-1:2025, chapter 8, ultimate limit state."""
+    """Validation for formula 8.51aw (hollow sections) from EN 1993-1-1:2022, chapter 8, ultimate limit state."""
 
     def test_evaluation(self) -> None:
         """Tests the evaluation of the result."""
@@ -71,7 +71,7 @@ class TestForm8Dot51awHollowSections:
 
 
 class TestForm8Dot51awWeldedBoxSections:
-    """Validation for formula 8.51aw (welded box sections) from EN 1993-1-1:2025, chapter 8, ultimate limit state."""
+    """Validation for formula 8.51aw (welded box sections) from EN 1993-1-1:2022, chapter 8, ultimate limit state."""
 
     def test_evaluation(self) -> None:
         """Tests the evaluation of the result."""
